HACK: sync Revert 4.9 changes of sync/fences to 4.4 state
CHROMIUM: android: fix warning when releasing active sync point
HACK: sync: Revert 4.6 changes to sync/fence
Revert "dma-buf/fence: make fence context 64 bit v2"
Revert "dma-buf/fence: add fence_array fences v6"
Revert "dma-buf/fence: add signal_on_any to the fence array v2"
Revert "dma-buf/sync_file: fix build warning with context format type"
Revert "staging/android: store last signaled value on sync timeline"
Revert "staging/android: remove .{fence, timeline}_value_str() from timeline_ops"
Revert "staging/android: remove struct sync_timeline_ops"
Revert "staging/android: remove sw_sync_timeline and sw_sync_pt"
Revert "staging/android: remove sw_sync.[ch] files"
Revert "staging/android: rename android_fence to timeline_fence"
Revert "staging/android: remove unnecessary check for fence"
Revert "staging/android: remove size arg of sync_timeline_create()"
Revert "staging/android: bring struct sync_pt back"
Revert "staging/android: move sw_sync related code to sw_sync.c"
Revert "staging/android: clean up #includes in the sync framework"
Revert "staging/android: make sync_timeline internal to sw_sync"
Revert "staging/android: make sw_ioctl info internal to sw_sync.c"
Revert "staging/android: remove 'destroyed' member from struct sync_timeline"
Revert "staging/android: remove sync_timeline_destroy()"
Revert "staging/android: remove drv_name from sync_timeline"
Revert "staging/android: rename sync.h to sync_debug.h"
Revert "staging/android: add DEBUG_FS dependence on Kconfig"
Revert "staging/android: sync_debug: unproxify debugfs files' fops"
Revert "dma-buf: propagate errors from dma_buf_describe() on debugfs read"
Revert "dma-buf: remove dma_buf directory on bufinfo file creation errors"
Revert "dma-buf: remove dma_buf_debugfs_create_file()"
Revert "dma-buf/sync_file: improve Kconfig description for Sync Files"
Revert "dma-buf: Release module reference on creation failure"
Revert "dma-buf/fence-array: add fence_is_array()"
Revert "dma-buf/sync_file: refactor fence storage in struct sync_file"
Revert "dma-buf/sync_file: add sync_file_get_fence()"
Revert "dma-buf/sync_file: only enable fence signalling on poll()"
Revert "staging/android: do not let userspace trigger WARN_ON"
Revert "staging/android: move trace/sync.h to sync_trace.h"
Revert "staging/android: prepare sw_sync files for de-staging"
Revert "staging/android: add Doc for SW_SYNC ioctl interface"
Revert "dma-buf/sw_sync: de-stage SW_SYNC"
Revert "dma-buf: Wait on the reservation object when sync'ing before CPU access"
Revert "dma-buf: fix kernel-doc warning and typos"
Revert "reservation: fix small comment typo"
Revert "dma-buf/sync-file: Avoid enable fence signaling if poll(.timeout=0)"
Revert "dma-buf/sync_file: Increment refcount of fence when all are signaled."
Revert "dma-buf/sync_file: free fences array in num_fences is 1"
Revert "CHROMIUM: fix warning when releasing active sync point"
Revert "uapi: add missing install of sync_file.h"
Revert "UPSTREAM: dma-buf: Introduce fence_get_rcu_safe()"
Revert "UPSTREAM: dma-buf: Restart reservation_object_get_fences_rcu() after writes"
Revert "UPSTREAM: dma-buf: Restart reservation_object_wait_timeout_rcu() after writes"
Revert "UPSTREAM: dma-buf: Restart reservation_object_test_signaled_rcu() after writes"
Revert "BACKPORT: dma-buf: Use fence_get_rcu_safe() for retrieving the exclusive fence"
Revert "BACKPORT: drm/atomic: add drm_atomic_set_fence_for_plane()"
Revert "BACKPORT: drm/fence: add in-fences support"
Revert "BACKPORT: drm/fence: add out-fences support"
Revert "UPSTREAM: drm/fences: add DOC: for explicit fencing"
Revert "BACKPORT: drm/fence: add drm_crtc_create_fence()"
Revert "BACKPORT: drm/fb_cma_helper: Add drm_fb_cma_prepare_fb() helper"

+++ drivers/gpu/drm/amd/amdgpu/amdgpu.h
@@ -2063,7 +2063,7 @@ struct amdgpu_device {
 	struct amdgpu_irq_src		hpd_irq;
 
 	/* rings */
-	u64				fence_context;
+	unsigned			fence_context;
 	unsigned			num_rings;
 	struct amdgpu_ring		*rings[AMDGPU_MAX_RINGS];
 	bool				ib_pool_ready;
